requirements.txt
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
chromadb>=0.5.0
rank-bm25>=0.2.2
tiktoken>=0.7.0
requests>=2.31.0
unidecode>=1.3.8
python-dotenv>=1.0.1
